pyproject.toml
[tool.poetry]
name = "django-projetointegrador"
version = "0.1.0"
description = "Backend do Projeto Integrador"
authors = ["Erico Pieri <erico.pieri@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
flake8 = "^5.0.4"
black = "^22.8.0"
django-cors-headers = "^3.13.0"
django-storages = {extras = ["azure"], version = "^1.13.1"}
boto3 = "^1.24.80"
Pillow = "^9.2.0"
python-decouple = "^3.6"
psycopg2 = "^2.9.4"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
drf-spectacular = "^0.24.2"

[tool.poetry.dev-dependencies]
black = {version = "^22.8.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
